Minimised to just a highlight colour selector
Updated docs to provide credit to creator.
