--- CHANGELOG.md
@@ -1,5 +1,0 @@
-# Changelog
-
-## [1.0.0] - 2025-08-15
-
-- Initial release
