+++ README.md
@@ -1,65 +1,18 @@
-# Swing Color Picker
+# Swing Minimal Color Picker
 
-A Java Swing color picker built with FlatLaf and MigLayout.
+A Java Swing color picker built with FlatLaf and MigLayout. Forked from https://github.com/DJ-Raven/swing-color-picker 
+and minimised to <300 LoC. Made the selected colour highlight persist
 
-<img src="screenshot/screenshot.jpg" alt="color-picker" width="600"/>&nbsp;
+![image info](./screenshot/darcula.png)
+![image info](./screenshot/intellij_light.png)
 
-## Installation
-
-[![Maven Central](https://img.shields.io/maven-central/v/io.github.dj-raven/swing-color-picker?label=Maven%20Central)](https://central.sonatype.com/artifact/io.github.dj-raven/swing-color-picker/versions)
-
-Add the dependency
-
-``` xml
-<dependency>
-    <groupId>io.github.dj-raven</groupId>
-    <artifactId>swing-color-picker</artifactId>
-    <version>1.0.0</version>
-</dependency>
-```
-
-### Snapshots
-
-To get the latest updates before the release, you can use the snapshot version
-from [Sonatype Central](https://central.sonatype.com/service/rest/repository/browse/maven-snapshots/io/github/dj-raven/swing-color-picker/)
-
-``` xml
-<repositories>
-    <repository>
-        <name>Central Portal Snapshots</name>
-        <id>central-portal-snapshots</id>
-        <url>https://central.sonatype.com/repository/maven-snapshots/</url>
-    </repository>
-</repositories>
-```
-
-Add the snapshot version
-
-``` xml
-<dependency>
-    <groupId>io.github.dj-raven</groupId>
-    <artifactId>swing-color-picker</artifactId>
-    <version>1.0.0-SNAPSHOT</version>
-</dependency>
-```
 
 ## Usage
-
-``` java
-// create color picker object
-ColorPicker colorPicker = new ColorPicker();
-
-// create event color changed
-colorPicker.addColorChangedListener((color, event) -> {
-    // color changed
-});
-
-```
 
 #### Show with Dialog
 
 ``` java
-Color color = ColorPicker.showDialog(this, "Pick Color", Color.WHITE);
+Color color = ColorPicker.showDialog(this, "Pick Highlight Color", new SimpleColorPicker());
 
 if (color != null) {
     // color selected
