--- src/main/java/raven/color/component/utils/RecentColorPaletteData.java
@@ -1,35 +1,0 @@
-package raven.color.component.utils;
-
-import java.awt.*;
-
-public class RecentColorPaletteData implements ColorPaletteData {
-
-    public RecentColorPaletteData() {
-        init();
-    }
-
-    private void init() {
-    }
-
-    @Override
-    public void add(Color color) {
-    }
-
-    @Override
-    public void remove(Color color) {
-    }
-
-    @Override
-    public void clear() {
-    }
-
-    @Override
-    public int size() {
-        return 0;
-    }
-
-    @Override
-    public Color get(int index) {
-        return null;
-    }
-}
