--- src/main/java/raven/color/event/ColorChangeEvent.java
@@ -1,17 +1,0 @@
-package raven.color.event;
-
-import java.util.EventObject;
-
-public class ColorChangeEvent extends EventObject {
-
-    private final boolean hueChanged;
-
-    public ColorChangeEvent(Object source, boolean hueChanged) {
-        super(source);
-        this.hueChanged = hueChanged;
-    }
-
-    public boolean isHueChanged() {
-        return hueChanged;
-    }
-}
