--- src/main/java/raven/color/event/ColorChangedListener.java
@@ -1,9 +1,0 @@
-package raven.color.event;
-
-import java.awt.*;
-import java.util.EventListener;
-
-public interface ColorChangedListener extends EventListener {
-
-    void colorChanged(Color color, ColorChangeEvent colorChangeEvent);
-}
